Move Subscription model and ROLE constants in a separate module.
