+++ opengever/activity/upgrades/to4503.py
@@ -1,6 +1,6 @@
-from opengever.activity.model.resource import TASK_ISSUER_ROLE
-from opengever.activity.model.resource import TASK_RESPONSIBLE_ROLE
-from opengever.activity.model.resource import WATCHER_ROLE
+from opengever.activity.model.subscription import TASK_ISSUER_ROLE
+from opengever.activity.model.subscription import TASK_RESPONSIBLE_ROLE
+from opengever.activity.model.subscription import WATCHER_ROLE
 from opengever.core.upgrade import SchemaMigration
 from sqlalchemy import Column
 from sqlalchemy import Integer
