+++ opengever/inbox/tests/test_activities.py
@@ -2,9 +2,9 @@ from ftw.builder import Builder
 from ftw.builder import create
 from ftw.testbrowser import browsing
 from opengever.activity import notification_center
-from opengever.activity.center import TASK_ISSUER_ROLE
-from opengever.activity.center import TASK_RESPONSIBLE_ROLE
 from opengever.activity.model import Activity
+from opengever.activity.model.subscription import TASK_ISSUER_ROLE
+from opengever.activity.model.subscription import TASK_RESPONSIBLE_ROLE
 from opengever.core.testing import OPENGEVER_FUNCTIONAL_ACTIVITY_LAYER
 from opengever.task.browser.accept.utils import accept_forwarding_with_successor
 from opengever.task.browser.accept.utils import assign_forwarding_to_dossier
