+++ opengever/task/browser/accept/utils.py
@@ -1,7 +1,7 @@
 from five import grok
 from opengever.activity import notification_center
-from opengever.activity.center import TASK_ISSUER_ROLE
-from opengever.activity.center import TASK_RESPONSIBLE_ROLE
+from opengever.activity.model.subscription import TASK_ISSUER_ROLE
+from opengever.activity.model.subscription import TASK_RESPONSIBLE_ROLE
 from opengever.base.oguid import Oguid
 from opengever.base.request import dispatch_request
 from opengever.base.transport import Transporter
