+++ opengever/task/tests/test_activities.py
@@ -4,10 +4,10 @@ from ftw.builder import create
 from ftw.testbrowser import browsing
 from ftw.testing.mailing import Mailing
 from opengever.activity import notification_center
-from opengever.activity.center import TASK_ISSUER_ROLE
-from opengever.activity.center import TASK_RESPONSIBLE_ROLE
-from opengever.activity.center import WATCHER_ROLE
 from opengever.activity.model import Activity
+from opengever.activity.model.subscription import TASK_ISSUER_ROLE
+from opengever.activity.model.subscription import TASK_RESPONSIBLE_ROLE
+from opengever.activity.model.subscription import WATCHER_ROLE
 from opengever.core.testing import OPENGEVER_FUNCTIONAL_ACTIVITY_LAYER
 from opengever.task.browser.accept.utils import accept_task_with_successor
 from opengever.testing import FunctionalTestCase
