+++ opengever/testing/builders/activity.py
@@ -1,11 +1,11 @@
 from ftw.builder import builder_registry
-from opengever.activity.center import WATCHER_ROLE
 from opengever.activity.model import Activity
 from opengever.activity.model import Notification
 from opengever.activity.model import NotificationDefault
 from opengever.activity.model import Resource
 from opengever.activity.model import Subscription
 from opengever.activity.model import Watcher
+from opengever.activity.model.subscription import WATCHER_ROLE
 from opengever.base.oguid import Oguid
 from opengever.ogds.models.tests.builders import SqlObjectBuilder
 from opengever.testing.builders.base import TEST_USER_ID
